Tidy up devservice/devmode related code in Hibernate extensions
