+++ extensions/hibernate-orm/deployment/src/test/java/io/quarkus/hibernate/orm/dev/H2CustomDialect.java
@@ -1,4 +1,4 @@
-package io.quarkus.hibernate.orm.devconsole;
+package io.quarkus.hibernate.orm.dev;
 
 import org.hibernate.dialect.H2Dialect;
 import org.hibernate.metamodel.mapping.EntityMappingType;
