+++ extensions/hibernate-orm/deployment/src/test/java/io/quarkus/hibernate/orm/dev/HibernateOrmDevControllerFailingDDLGenerationTestCase.java
@@ -1,4 +1,4 @@
-package io.quarkus.hibernate.orm.devconsole;
+package io.quarkus.hibernate.orm.dev;
 
 import static org.hamcrest.Matchers.is;
 
@@ -8,21 +8,23 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import io.quarkus.test.QuarkusDevModeTest;
 import io.restassured.RestAssured;
 
-public class DevConsolePuInformationTestCase {
+public class HibernateOrmDevControllerFailingDDLGenerationTestCase {
     @RegisterExtension
     final static QuarkusDevModeTest TEST = new QuarkusDevModeTest()
             .withApplicationRoot((jar) -> jar
-                    .addClasses(MyEntityWithSuccessfulDDLGeneration.class,
-                            DevConsoleInfoSupplierTestResource.class)
-                    .addAsResource("application.properties")
+                    .addClasses(MyEntityWithFailingDDLGeneration.class,
+                            TypeWithUnsupportedSqlCode.class,
+                            H2CustomDialect.class,
+                            HibernateOrmDevInfoServiceTestResource.class)
+                    .addAsResource("application-generation-none-customh2.properties", "application.properties")
                     .addAsResource("import-custom-table-name.sql", "import.sql"));
 
     @Test
-    public void infoAvailable() {
+    public void infoAvailableButWithException() {
         RestAssured.given()
                 .param("expectedCreateDDLContent", "EXCEPTION")
                 .param("expectedDropDDLContent", "EXCEPTION")
-                .when().get("/dev-console-info-supplier/check-pu-info-with-successful-ddl-generation")
+                .when().get("/dev-info/check-pu-info-with-failing-ddl-generation")
                 .then().body(is("OK"));
     }
