+++ extensions/hibernate-orm/deployment/src/test/java/io/quarkus/hibernate/orm/dev/HibernateOrmDevControllerTestCase.java
@@ -1,4 +1,4 @@
-package io.quarkus.hibernate.orm.devconsole;
+package io.quarkus.hibernate.orm.dev;
 
 import static org.hamcrest.Matchers.is;
 
@@ -8,23 +8,21 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import io.quarkus.test.QuarkusDevModeTest;
 import io.restassured.RestAssured;
 
-public class DevConsolePuInformationFailingDDLGenerationTestCase {
+public class HibernateOrmDevControllerTestCase {
     @RegisterExtension
     final static QuarkusDevModeTest TEST = new QuarkusDevModeTest()
             .withApplicationRoot((jar) -> jar
-                    .addClasses(MyEntityWithFailingDDLGeneration.class,
-                            TypeWithUnsupportedSqlCode.class,
-                            H2CustomDialect.class,
-                            DevConsoleInfoSupplierTestResource.class)
-                    .addAsResource("application-generation-none-customh2.properties", "application.properties")
+                    .addClasses(MyEntityWithSuccessfulDDLGeneration.class,
+                            HibernateOrmDevInfoServiceTestResource.class)
+                    .addAsResource("application.properties")
                     .addAsResource("import-custom-table-name.sql", "import.sql"));
 
     @Test
-    public void infoAvailableButWithException() {
+    public void infoAvailable() {
         RestAssured.given()
                 .param("expectedCreateDDLContent", "EXCEPTION")
                 .param("expectedDropDDLContent", "EXCEPTION")
-                .when().get("/dev-console-info-supplier/check-pu-info-with-failing-ddl-generation")
+                .when().get("/dev-info/check-pu-info-with-successful-ddl-generation")
                 .then().body(is("OK"));
     }
