+++ extensions/hibernate-orm/deployment/src/test/java/io/quarkus/hibernate/orm/dev/TypeWithUnsupportedSqlCode.java
@@ -1,4 +1,4 @@
-package io.quarkus.hibernate.orm.devconsole;
+package io.quarkus.hibernate.orm.dev;
 
 import java.io.Serializable;
 import java.sql.PreparedStatement;
