+++ extensions/hibernate-orm/deployment/src/test/resources/application-generation-none-customh2.properties
@@ -4,4 +4,4 @@ quarkus.datasource.jdbc.url=jdbc:h2:mem:test
 #quarkus.hibernate-orm.log.sql=true
 quarkus.hibernate-orm.database.generation=none
 #We need the custom dialect to force the failure to happen exclusively during DDL generation
-quarkus.hibernate-orm.dialect=io.quarkus.hibernate.orm.devconsole.H2CustomDialect
+quarkus.hibernate-orm.dialect=io.quarkus.hibernate.orm.dev.H2CustomDialect
