+++ extensions/hibernate-orm/runtime/src/main/java/io/quarkus/hibernate/orm/runtime/dev/HibernateOrmDevInfoCreateDDLSupplier.java
@@ -1,4 +1,4 @@
-package io.quarkus.hibernate.orm.runtime.devconsole;
+package io.quarkus.hibernate.orm.runtime.dev;
 
 import java.util.Collection;
 import java.util.Objects;
@@ -6,19 +6,19 @@ import java.util.function.Supplier;
 
 import io.quarkus.runtime.annotations.RecordableConstructor;
 
-public class HibernateOrmDevConsoleCreateDDLSupplier implements Supplier<String> {
+public class HibernateOrmDevInfoCreateDDLSupplier implements Supplier<String> {
 
     private final String puName;
 
     @RecordableConstructor
-    public HibernateOrmDevConsoleCreateDDLSupplier(String puName) {
+    public HibernateOrmDevInfoCreateDDLSupplier(String puName) {
         this.puName = puName;
     }
 
     @Override
     public String get() {
-        Collection<HibernateOrmDevConsoleInfoSupplier.PersistenceUnitInfo> persistenceUnits = HibernateOrmDevConsoleInfoSupplier.INSTANCE
-                .getPersistenceUnits();
+        Collection<HibernateOrmDevInfo.PersistenceUnit> persistenceUnits = HibernateOrmDevController.get()
+                .getInfo().getPersistenceUnits();
         for (var p : persistenceUnits) {
             if (Objects.equals(puName, p.getName())) {
                 return p.getCreateDDL();
