+++ extensions/hibernate-orm/runtime/src/main/java/io/quarkus/hibernate/orm/runtime/dev/HibernateOrmDevInfoSupplier.java
@@ -1,0 +1,12 @@
+package io.quarkus.hibernate.orm.runtime.dev;
+
+import java.util.function.Supplier;
+
+public class HibernateOrmDevInfoSupplier implements Supplier<HibernateOrmDevInfo> {
+
+    @Override
+    public HibernateOrmDevInfo get() {
+        return HibernateOrmDevController.get().getInfo();
+    }
+
+}
