+++ extensions/hibernate-search-orm-elasticsearch/runtime/src/main/java/io/quarkus/hibernate/search/orm/elasticsearch/runtime/dev/HibernateSearchElasticsearchDevInfoSupplier.java
@@ -1,0 +1,13 @@
+package io.quarkus.hibernate.search.orm.elasticsearch.runtime.dev;
+
+import java.util.function.Supplier;
+
+public class HibernateSearchElasticsearchDevInfoSupplier
+        implements Supplier<HibernateSearchElasticsearchDevInfo> {
+
+    @Override
+    public HibernateSearchElasticsearchDevInfo get() {
+        return HibernateSearchElasticsearchDevController.get().getInfo();
+    }
+
+}
